+++ docs/CHANGELOG.md
@@ -7,6 +7,17 @@ Alle viktige endringer i NISSY-scriptene vil bli dokumentert i denne filen.
 - Ser på mulighetene for enkel samkjøringsalgoritme
 
 ---
+## 🚀 [3.2.0] - ??.01.2026
+
+### Masse nytt og forbedringer til Avbestilling-script!
+
+#### Hva er nytt?
+- ✖️ **Avbestilling** 
+  - Forbedret utgråingslogikk ved avbestilling av turer/bestillinger, bruker NISSY sin egen funksjon riktig.
+  - Kan nå bruke kryss på ventende, pågående og i ressurser og få ny forbedret pop-up
+  - Sperrer av muligheten for å avbestille ved kryss etter 3003 XML.
+  - Gir advarsler til bruker ved avplanlegging av samkjørte bestillinger
+
 ## 🚀 [3.1.0] - 23.01.2026
 
 ### Forbedring til Smart-tildel